Add a hack to ignore DW_AT_bit_size for pointers.
This appears to be a regression in the recent compiler upgrade as clang
is now generating DW_AT_bit_size attributes for pointers which are
members of a struct.  GDB interprets DW_AT_bit_size to mean that these
pointers are bitfields instead of pointers.

Use 'check_typedef()' to resolve the type through any typedefs to the
root type to determine if it is really a pointer.