First cut at a manual stack unwinder for CheriABI.
As with the other manual MIPS unwinders, this unwinder works by scanning
a function prologue and decoding instructions to figure out which registers
are saved on the stack and where.

As part of the unwinder, add a CheriABI version of the routine to find the
end of a previous function by looking for CHERI branches (cjr, cjalr).

This appears to help with unwinding out of syscall stubs, perhaps just due
to finding the end of the previous function correctly.