Add support for user-only qtrace
New commands `qtrace start-user` and `qtrace start-full` can be used to
select which mode of tracing is used. A cleaner solution would be if there
was a ptrace option and a flag `MDTD_QTRACE_USER_ONLY` but that adds even
more overhead to process switching and should not be necessary as long as
the debugged process doesn't use the magic nop.

Fixes #3