Try to use PCC more often for unwinding PC if PCC exists.
Notably, for hybrid binaries, PC is not an abolute virtual address if
PCC has a non-zero base.  To handle this, merge mips_fbsd_cheri_read_pc
into mips_read_pc.  To read the top-level PC value (not the register,
but the logical instruction pointer), always read PCC's cursor if we
have a PC.  For the gdbarch_unwind_pc method, leave a CheriAB-specific
method in mips-fbsd-tdep.c, but modify the mips_unwind_pc() method to
look for PCC and if it can find a PCC value add its base to the value
of $pc to compute the effective PC.

Note that the current attempt to fetch PCC might fail if PCC isn't
saved in later frames down the stack in which case this will have to
be revisited.

Fix #1.