Better handling of the magic mode bit in CLR.
Borrowing from how the ARM bits in GDB handle Thumb mode, add
support for unwinding CPSR, and at least faking up the PSTATE.C64
bit by looking at the low bit of the saved PCC in CLR.

Clear the low bit from PCC when unwinding it from CLR instead of
just copying CLR into PCC.  This fixes the PC values in backtraces
to be the real instructions and not off by one.

Recognize function symbols with an odd address and mark them as
"special" (which means they execute in capability mode) and update the
raw address to be addr - 1.  This moves the address of the symbol back
to the real starting address of the function, and means that
"disassemble" now works.