[CHERI-RISC-V] Support cap-mode dependent disassembly
With this change mode-dependent instructions such as AUIPCC/CLC/etc. will
be disassembled as the capmode version.
In order to do determine the mode we are in right now we pass an additional
flag to the disassembler. If this flag is not passed (e.g. in the objdump
case), we fall back to querying the ELF header to check for EF_RISCV_CAPMODE.

Note: this is currently not quite correct for GDB: there is no cached value
for capmode so we use is_cheriabi instead (which is generally the same but
is not necessarily true).

TODO: A follow-up change should set the value of capmode based on PCC.flags
when the current PCC is available instead of statically based on the ELF
header of the main program.