Support for disassembling CHERI-RISC-V capmode instructions.
Add a separate table of capmode instructions and change
riscv_disassemble_insn to look for a match in this table first when
using capmode.  If a match isn't found, fall back to the normal table.

Capmode is currently enabled for any ELF file which has the capmode
flag set in the ELF header flags.