gdb tdesc: Handle mismatched pointer register types.
In the XML target descriptions, registers can be given a type of
"code_ptr" or "data_ptr".  GDB always uses the builtin type for "void
*" for these registers.  However, this is wrong if the register's size
does not match (e.g. the legacy "sp" and "pc" ARM64 registers when
using a purecap binary for which "void *" is a capability).  If the
sizes don't match, try to find a matching type such as "long" or
"intcap_t" to use instead.