morello: Basic (possibly wrong) support for restricted mode unwinding.
For purecap frames, check the executive permission in the saved CLR of
the next frame (which is the PCC at the time of calling down to the
next frame) to determine if the frame is in executive or restricted
mode.  Use this to decide which register pulls from CFA during dwarf
unwinding as well as to decide which register's address to return for
gdbarch_unwind_sp.