riscv: Add a gdbarch_write_pc hook for purecap.
This tries to derive a new PCC from the existing value of PCC and the
requested address.  If the new address is not representable, the tag
of PCC is cleared.  A more correct (but much more complicated)
solution would be to walk the list of shared libraries choosing bounds
for the address if it is not in bounds of the current PCC.

However, this is sufficient for 'start' to work once PT_SETCAPREGS is
enabled.