morello: Rework morello_write_pc.
If the new PC is the same as the old, just leave it unchanged.

At the start of run, it calls proceed() with the current PC and
proceed() turns around and writes that PC back to the regcache.  The
current version of morello_write_pc discarded the bounds during this
round trip.

While here, make morello_write_pc a bit closer to RISC-V in that it
tries to use the existing bounds if the new PC is in bounds.

Unlike RISC-V this retains the previous behavior of falling back to
broad bounds if the PC would be untagged otherwise.