morello_write_pc: Use derive_capability_for_address
Read the current PCC as a struct value instead of a raw buffer to
permit using derive_capability_for_address.  Still try to derive from
the current PCC first if possible.  However, if that fails, fall back
to trying to derive a valid PCC rather than the infinite bounds
fallback.