compart: Add initial compartment support
This adds a new 'info compartments' command which lists the current
compartments similar to 'info sharedlibrary' for shared libraries.

This depends on two new gdbarch hooks: one to fetch the current list
of compartments for the current process and one to fetch detailed info
for each compartment.  The first call merely needs to fetch the name,
ID, and address of the compartment structure.  When fetching a new
list of compartments, if the name, ID, and address are the same, GDB
assumes the compartment hasn't changed and caches it.  Otherwise when
a compartment is first added the second hook is used to fetch more
detailed information like the list of libraries.