symfile: Use compartment-specific sections as default object file sections
In various places GDB wants to have a definitive .text section (and a
few others).  For example, when parsing high and low PC values in
DWARF DIEs, GDB wants to relocate those values to virtual addresses by
using the relative relocation offset for the .text section.  However,
binaries and shared object files with compartments may not have those
sections in the default compartment.  Instead, find a suitable section
from any compartment to use as the representative section.

Possibly the DWARF reader should be looking up the relevant section
containing each unrelocated address and relocating the address using
the offset of the containing section.  However, that is more work, and
I'm worried there are other places that would need similar fixes.
This is a simpler change for now, and is fine in practice as all
sections are relocated relative to the same base address anyway.