value_from_component: Handle caps from values without backing store
Since values like structs and arrays do not currently store tags of
member objects (just the 16 bytes), value_from_component always
allocates lazy values for capability components forcing the full
capability to be fetched from the backing store.  This works fine
(albeit a bit inefficient) for values backed by memory or registers.
It does not work for synthetic objects created as a result of
evaluating expressions, and in particular arrays synthesized by GDB's
expression parser.  For this case, alow the 16 bytes to be copied from
the containing object, but explicitly mark the value as having an
invalid tag.

The real fix for this is to stop storing a single tag as sideband data
in struct value, but to instead embed tags in the raw array of bytes
as a 17th byte.  However, that is a pretty far-ranging fix.  This
band-aid will at least prevent crashes for now at the expense of lying
about component members that are actually tagged.