fixup! aarch64-fbsd-tdep: Add unwinder for non-benchmark c18n trampolines
CLR is the second field, not the third, which is C19. Most of the time
the transitions are from non-executive, and unless C19 happens to
contain a function pointer it will be non-executive, so this worked most
of the time, but fails when signal handlers are involved, in particular.

Moreover, just as we restore the X registers above, we also need to
restore the integer SP (no integer equivalent of RCSP/ECSP exists),
otherwise it will be out of sync with CSP, and any attempt to unwind
further using SP will use the wrong address; this is possibly more
important for -O0 code, and also helps when debug info is not available.